fix: int64 cache-offset arithmetic in sparse Triton kernels (long-context OOB)
ROOT CAUSE of the intermittent cudaErrorIllegalAddress at long context /
bsz>1: every sparse kernel computed its KV-cache byte offset as

    block_id * cache_stride_block + ...

with block_id loaded from the block table as int32 and cache_stride_block an
int32 kernel arg, so the product was evaluated in int32. The physical block id
is an index into the *whole* paged cache, which on a B200 (183 GB) fills to
millions of blocks. For an LRoSA slot (2*head+cs_h => cache_stride_block=36864)
int32 overflows once block_id exceeds ~58k. Whether the wrapped (negative)
offset lands on an unmapped page depends on the caching allocator's layout, so
the fault looked intermittent — surfacing under memory pressure (profiler,
concurrent runs, long context where high block ids are actually reached) and
vanishing under CUDA_LAUNCH_BLOCKING=1 / short context. It was NOT a CUDA-graph
or radix-vs-topk issue: a torch.topk bisect crashed identically because the
shared score kernel has the same arithmetic.

Fix: cast block_id to int64 at every block-table / page-block load so the whole
offset expression promotes to int64. Sites:
  - triton_lrosa_score_topk.py: per-kv-head + per-layer-concat score kernels
  - triton_lrosa_gather.py: K/V selection gather
  - triton_lrosa_streaming_topk.py: streaming score kernel
  - triton_fasa_score_topk.py: FASA-fc channel-gather score kernel
  - triton_quest.py: all 7 cache-indexing kernels (repr store, page min/max
    score, trailing/selected/streamed attention, K/V gather) — same latent bug,
    expected to also fix Quest bsz>1.

Also harden _radix_topk: rewrite selected indices that are negative (the -1 pad)
OR >= seq_len to a real token in [0, seq_len), so the downstream gather can
never see an out-of-range index regardless of what the radix kernel returns
(defensive; the int64 fix removes the actual fault).

Verified: LRoSA Qwen3-8B prefill=64k bsz=8 decode now runs clean under the
torch profiler (previously raised cudaErrorIllegalAddress mid-decode).

+++ vllm/v1/attention/ops/triton_fasa_score_topk.py
@@ -82,7 +82,9 @@ def _fasa_score_kernel(
     block_idx = t_offs // block_size
     pos_in_block = t_offs % block_size
     bt_offset = pid_r * bt_stride_r + block_idx
-    block_id = tl.load(block_table_ptr + bt_offset, mask=in_seq, other=0)
+    # int64: physical block ids overflow int32 cache-offset arithmetic once the
+    # KV cache fills large memory -> intermittent OOB at long context.
+    block_id = tl.load(block_table_ptr + bt_offset, mask=in_seq, other=0).to(tl.int64)
 
     # K region starts at offset 0 within the slot; gather the I_dom channels.
     K_base = (
